requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
Faker==25.3.0
idna==3.7
iniconfig==2.0.0
mylib==0.0.1
packaging==24.0
pluggy==1.5.0
pytest==8.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
sentry-sdk==2.4.0
six==1.16.0
urllib3==2.2.1
